requirements.txt
scapy-http==1.8.2
pandas
torch
sklearn
matplotlib
